requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dj-stripe==2.8.4
Django==5.0.4
django-redis==5.4.0
djangorestframework==3.15.1
idna==3.7
iniconfig==2.0.0
kafka==1.3.5
packaging==24.0
pluggy==1.4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pytest==8.1.1
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sqlparse==0.4.4
stripe==9.1.0
typing_extensions==4.11.0
urllib3==2.2.1
